updated conversion of size to string
git-svn-id: https://eriforce.eicp.net:8443/svn/PureLib@96 4dedb8ab-72de-3345-9651-2401c5e3755d

+++ trunk/PureLib/Common/UnitDisplay.cs
@@ -40,7 +40,7 @@ namespace PureLib.Common {
             return (long)result;
         }
 
-        public static string ToFriendlyString(this decimal size, int digits = 2, Dictionary<SizeUnit, string> unitNames = null) {
+        public static string ToFriendlyString(this long size, int digits = 2, Dictionary<SizeUnit, string> unitNames = null) {
             int maxIndex = Enum.GetValues(typeof(SizeUnit)).Length - 1;
             int unitIndex = 0;
             while ((size >= 1000) && (unitIndex < maxIndex)) {
@@ -48,7 +48,7 @@ namespace PureLib.Common {
                 unitIndex++;
             }
             SizeUnit unit = (SizeUnit)unitIndex;
-            return pairFormat.FormatWith(Math.Round(size, digits), unit.ToDisplayName(unitNames));
+            return pairFormat.FormatWith(Math.Round((decimal)size, digits), unit.ToDisplayName(unitNames));
         }
 
         public static string ToFriendlyString(this TimeSpan ts, DateTimeUnit? truncateUnit = null,
